Get rid of bft_test_utils
