+++ Cargo.toml
@@ -14,12 +14,11 @@ env_logger = "0.11.3"
 log = "0.4.21"
 
 [workspace]
-members = ["bft_interp", "bft_types", "bft_test_utils"]
+members = ["bft_interp", "bft_types"]
 
 [dev-dependencies]
 criterion = "0.5.1"
 tempfile = "3.10.0"
-bft_test_utils = { path = "./bft_test_utils" }
 
 [[bench]]
 name = "benchmarks"
