+++ bft_interp/Cargo.toml
@@ -13,5 +13,5 @@ log = "0.4.21"
 env_logger = "0.11.3"
 
 [dev-dependencies]
-bft_test_utils = { path = "../bft_test_utils" }
 rand = "0.8.5"
+tempfile = "3.10.0"
