--- bft_test_utils/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "bft_test_utils"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tempfile = "3.10.0"
-log = "0.4.21"
-env_logger = "0.11.3"
-
-[dev-dependencies]
-bft_types = { path = "../bft_types" }
-bft_interp = { path = "../bft_interp" }
