+++ bft_types/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dev-dependencies]
-bft_test_utils = { path = "../bft_test_utils" }
+tempfile = "3.10.0"
 
 [dependencies]
 num-traits = "0.2.18"
